+++ x-uploader/test/x/uploader/media_request_test.rb
@@ -9,7 +9,6 @@ module X
     cover Uploader::Chunks
 
     UPLOAD_URL = "https://api.x.com/2/media/upload".freeze
-    BOUNDARY = "AaB03x".freeze
     CONTENT = "\x89PNG\r\n\x1A\n\x00\x00\x00...".b.freeze
     GIF_FILE = "test/sample_files/sample.gif".freeze
     HEX_BOUNDARY = %r{\Amultipart/form-data; boundary=(\h{32})\z}
@@ -23,28 +22,23 @@ module X
     end
 
     def test_upload_binary_body_and_headers
-      Uploader::Media.upload_binary(CONTENT, client: @client, media_category: "tweet_image", boundary: BOUNDARY)
-
-      assert_equal "multipart/form-data; boundary=#{BOUNDARY}", @request.headers["Content-Type"]
-      assert_equal upload_body(CONTENT, "tweet_image", BOUNDARY), @request.body.b
-    end
-
-    def test_upload_sends_file_content
-      Uploader::Media.upload(GIF_FILE, client: @client, media_category: "tweet_gif", boundary: BOUNDARY)
-
-      assert_equal upload_body(File.binread(GIF_FILE), "tweet_gif", BOUNDARY), @request.body.b
-    end
-
-    def test_upload_binary_default_boundary
       Uploader::Media.upload_binary(CONTENT, client: @client, media_category: "tweet_image")
 
       assert_equal upload_body(CONTENT, "tweet_image", request_boundary), @request.body.b
     end
 
-    def test_upload_default_boundary
+    def test_upload_sends_file_content
       Uploader::Media.upload(GIF_FILE, client: @client, media_category: "tweet_gif")
 
       assert_equal upload_body(File.binread(GIF_FILE), "tweet_gif", request_boundary), @request.body.b
+    end
+
+    def test_each_upload_has_a_boundary_of_its_own
+      Uploader::Media.upload_binary(CONTENT, client: @client, media_category: "tweet_image")
+      first = request_boundary
+      Uploader::Media.upload_binary(CONTENT, client: @client, media_category: "tweet_image")
+
+      refute_equal first, request_boundary
     end
 
     def test_upload_binary_rejects_invalid_category_before_requesting